+++ rotkehlchen/db/upgrade_manager.py
@@ -36,6 +36,7 @@ from rotkehlchen.db.upgrades.v49_v50 import upgrade_v49_to_v50
 from rotkehlchen.db.upgrades.v50_v51 import upgrade_v50_to_v51
 from rotkehlchen.db.upgrades.v51_v52 import upgrade_v51_to_v52
 from rotkehlchen.db.upgrades.v52_v53 import upgrade_v52_to_v53
+from rotkehlchen.db.upgrades.v53_v54 import upgrade_v53_to_v54
 from rotkehlchen.errors.misc import DBUpgradeError
 from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.utils.misc import ts_now
@@ -78,6 +79,7 @@ UPGRADES_LIST = [
     UpgradeRecord(from_version=50, function=upgrade_v50_to_v51),
     UpgradeRecord(from_version=51, function=upgrade_v51_to_v52),
     UpgradeRecord(from_version=52, function=upgrade_v52_to_v53),
+    UpgradeRecord(from_version=53, function=upgrade_v53_to_v54),
 ]