+++ rotkehlchen/tests/unit/test_chains_aggregator.py
@@ -102,6 +102,7 @@ def test_detect_evm_accounts(blockchain: ChainsAggregator) -> None:
         (SupportedBlockchain.SCROLL, everywhere_addy, None),
         (SupportedBlockchain.BINANCE_SC, everywhere_addy, None),
         (SupportedBlockchain.MONAD, everywhere_addy, None),
+          (SupportedBlockchain.SONIC, everywhere_addy, None),
     ]
 
     for chain, addy, label in addies_to_start_with:
@@ -139,6 +140,7 @@ def test_detect_evm_accounts(blockchain: ChainsAggregator) -> None:
             scroll_addresses=[everywhere_addy, addy_eoa_3],
             binance_sc_addresses=[everywhere_addy, addy_eoa_3],
             monad_addresses=[everywhere_addy, addy_eoa_3],
+              sonic_addresses=[everywhere_addy, addy_eoa_3],
         )
 
         blockchain.detect_evm_accounts()
@@ -153,6 +155,7 @@ def test_detect_evm_accounts(blockchain: ChainsAggregator) -> None:
     assert set(blockchain.accounts.scroll) == {addy_eoa_3, everywhere_addy}
     assert set(blockchain.accounts.binance_sc) == {addy_eoa_3, everywhere_addy}
     assert set(blockchain.accounts.monad) == {addy_eoa_3, everywhere_addy}
+    assert set(blockchain.accounts.sonic) == {addy_eoa_3, everywhere_addy}
 
     # Also check the db
     expected_accounts_data = initial_accounts_data + [
@@ -171,7 +174,8 @@ def test_detect_evm_accounts(blockchain: ChainsAggregator) -> None:
             (SupportedBlockchain.SCROLL, addy_eoa_3, label2),
             (SupportedBlockchain.BINANCE_SC, addy_eoa_3, label2),
             (SupportedBlockchain.MONAD, addy_eoa_3, label2),
-        )
+              (SupportedBlockchain.SONIC, addy_eoa_3, label2),
+          )
     ]
 
     with blockchain.database.conn.read_ctx() as cursor:
@@ -197,6 +201,7 @@ def test_detect_evm_accounts(blockchain: ChainsAggregator) -> None:
             SupportedBlockchain.SCROLL,
             SupportedBlockchain.BINANCE_SC,
             SupportedBlockchain.MONAD,
+              SupportedBlockchain.SONIC,
         ) for account in raw_accounts.get(chain)
     ]